Create grid logic
Created js function that creates a 16x16 grid of divs with each div
containing text content of "hi" to test. Added css to flex the class
.gridRow so that div elements would space evenly around the page and
create the grid, but did not add any additonal code to ensure grid
objects are square. Tested in browser and verified to function on page
load.